+++ src/Framework/Helpers/ConversationTopicHelper.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using StardewModdingAPI;
+using StardewValley;
+
+
+namespace QuestFramework.Framework.Helpers
+{
+    internal class ConversationTopicHelper
+    {
+        internal protected static IMonitor Monitor => QuestFrameworkMod.Instance.Monitor;
+        static public void AddConversationTopic(string addConversationTopicInput)
+        {
+            string[] convTopicToAddParts = addConversationTopicInput.Split(' ');
+
+            if (convTopicToAddParts.Length % 2 == 0 && convTopicToAddParts.Length >= 2)
+            {
+                for (int i = 0; i < convTopicToAddParts.Length; i += 2)
+                {
+                    string convTopicToAdd = convTopicToAddParts[i];
+                    int convTopicCompletedDaysActive = Convert.ToInt32(convTopicToAddParts[i + 1]);
+
+                    if (Game1.player.activeDialogueEvents.ContainsKey(convTopicToAdd))
+                        continue;
+
+                    Game1.player.activeDialogueEvents.Add(convTopicToAdd, convTopicCompletedDaysActive);
+                    Monitor.Log($"Added conversation topic with the key: `{convTopicToAdd}`" +
+                        $"the conversation topic will be active for `{convTopicCompletedDaysActive}` days");
+                }
+            }
+
+        }
+        static public void RemoveConversationTopic(string removeConversationTopicInput)
+        {
+            foreach (string convTopicToRemove in removeConversationTopicInput.Split(' ') )
+            {
+                if (!Game1.player.activeDialogueEvents.ContainsKey(convTopicToRemove))
+                    continue;
+
+                Game1.player.activeDialogueEvents.Remove(convTopicToRemove);
+                Monitor.Log($"Removed conversation topic with the key: `{convTopicToRemove}`");        
+            }
+        }
+    }
+}